refactor(#1406): move cancelExistingSeed to seed-admission, split race test
Why:
- Maintainer review re-flagged F128 AC-X1: all files must stay at or below
  350 lines. proposal-approve-dispatch.ts (366) and
  proposal-approve-reconcile.test.js (391) were over the cap.

What changed:
- Moved cancelExistingSeed from proposal-approve-dispatch.ts into
  proposal-seed-admission.ts (focused terminal-seed transition module).
- Import it from proposal-approve-dispatch.ts; behavior unchanged.
- Split the orphan-carrier race test into a new focused module:
  proposal-race-rollback.test.js.

Verification:
- pnpm --filter @cat-cafe/api build passes.
- node --test on the 6 affected proposal test modules: 37 tests pass.
- biome check --write passes (only pre-existing cognitive complexity warnings
  in appendApprovedInitialMessage / executeQueuedDispatch remain).
